requirements.txt
openai>=1.30.0
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.1.0
numpy>=1.24.0
yfinance>=0.2.20
python-dateutil>=2.8.2
